[refactor] : producer와 config 구현  및 필요 없는 Kafka 관련 클래스 삭제 (#42)
